Improve check formatting for summary. Add extra context
