+++ src/knowledge-base/knowledgeBase.ts
@@ -10,3 +10,10 @@ export type KnowledgeDocument = {
 export const getAllDocuments = async (): Promise<KnowledgeDocument[]> => {
   return fetchDocumentsFromKnowledgeBase();
 };
+
+export const getDocumentBySlug = async (
+  slug: string,
+): Promise<KnowledgeDocument | null> => {
+  const documents = await getAllDocuments();
+  return documents.find((doc) => doc.slug === slug) || null;
+};
